Move forge installation to the forge folder
