+++ forge/.vscode/launch.json
@@ -6,7 +6,7 @@
       "name": "runClient",
       "request": "launch",
       "mainClass": "net.minecraftforge.bootstrap.ForgeBootstrap",
-      "projectName": "SpiderDens",
+      "projectName": "forge",
       "cwd": "${workspaceFolder}\\run",
       "vmArgs": "-Dforge.logging.console.level\u003ddebug -Dforge.logging.markers\u003dREGISTRIES -Dforge.enabledGameTestNamespaces\u003dspiderdens -Dforge.enableGameTest\u003dtrue -Djava.net.preferIPv6Addresses\u003dsystem -XX:HeapDumpPath\u003dMojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump",
       "args": "--launchTarget forge_userdev_client --version MOD_DEV --assetIndex 16 --assetsDir C:\\Users\\Paul\\.gradle\\caches\\forge_gradle\\assets --gameDir .",
@@ -21,10 +21,10 @@
       "name": "runData",
       "request": "launch",
       "mainClass": "net.minecraftforge.bootstrap.ForgeBootstrap",
-      "projectName": "SpiderDens",
+      "projectName": "forge",
       "cwd": "${workspaceFolder}\\run-data",
       "vmArgs": "-Dforge.logging.console.level\u003ddebug -Dforge.logging.markers\u003dREGISTRIES -Djava.net.preferIPv6Addresses\u003dsystem",
-      "args": "--launchTarget forge_userdev_data --assetIndex 16 --assetsDir C:\\Users\\Paul\\.gradle\\caches\\forge_gradle\\assets --gameDir . --mod spiderdens --all --output \"C:\\Users\\Paul\\OneDrive\\Documents\\DEV\\DEV Files\\SpiderDens\\SpiderDens\\src\\generated\\resources\" --existing \"C:\\Users\\Paul\\OneDrive\\Documents\\DEV\\DEV Files\\SpiderDens\\SpiderDens\\src\\main\\resources\"",
+      "args": "--launchTarget forge_userdev_data --assetIndex 16 --assetsDir C:\\Users\\Paul\\.gradle\\caches\\forge_gradle\\assets --gameDir . --mod spiderdens --all --output \"C:\\Users\\Paul\\OneDrive\\Documents\\DEV\\DEV Files\\SpiderDens\\SpiderDens\\forge\\src\\generated\\resources\" --existing \"C:\\Users\\Paul\\OneDrive\\Documents\\DEV\\DEV Files\\SpiderDens\\SpiderDens\\forge\\src\\main\\resources\"",
       "env": {
         "MOD_CLASSES": "${workspaceFolder}\\build\\sourcesSets\\main",
         "MCP_MAPPINGS": "official_1.20.6"
@@ -36,7 +36,7 @@
       "name": "runGameTestServer",
       "request": "launch",
       "mainClass": "net.minecraftforge.bootstrap.ForgeBootstrap",
-      "projectName": "SpiderDens",
+      "projectName": "forge",
       "cwd": "${workspaceFolder}\\run",
       "vmArgs": "-Dforge.logging.console.level\u003ddebug -Dforge.logging.markers\u003dREGISTRIES -Dforge.enabledGameTestNamespaces\u003dspiderdens -Djava.net.preferIPv6Addresses\u003dsystem",
       "args": "--launchTarget forge_userdev_server_gametest --gameDir .",
@@ -51,7 +51,7 @@
       "name": "runServer",
       "request": "launch",
       "mainClass": "net.minecraftforge.bootstrap.ForgeBootstrap",
-      "projectName": "SpiderDens",
+      "projectName": "forge",
       "cwd": "${workspaceFolder}\\run",
       "vmArgs": "-Dforge.logging.console.level\u003ddebug -Dforge.logging.markers\u003dREGISTRIES -Dforge.enabledGameTestNamespaces\u003dspiderdens -Dforge.enableGameTest\u003dtrue -Djava.net.preferIPv6Addresses\u003dsystem",
       "args": "--launchTarget forge_userdev_server --gameDir . --nogui",
