+++ forge/src/main/java/com/goopey/spiderdens/Config.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package com.goopey.spiderdens;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
