+++ forge/src/main/java/com/goopey/spiderdens/SpiderDens.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package com.goopey.spiderdens;
 
 import com.mojang.logging.LogUtils;
 import net.minecraft.client.Minecraft;
